Renaming django-favorite to django-faveit.
