+++ favit/static/css/favorite.css
