+++ favit/static/js/favorite.js
