+++ favit/templates/favit/button.html
@@ -1,5 +1,5 @@
 <div class="favorite">
-  <button class="btn favorite" href="{% url favorite.views.add_or_remove %}" model="{{ target_model }}" id="target_{{ target_object_id }}">
+  <button class="btn favorite" href="{% url favit.views.add_or_remove %}" model="{{ target_model }}" id="target_{{ target_object_id }}">
     <i class="icon-star{% if not undo %}-empty{% endif %}"></i>
   </button>
   <span class="fav-count">{{ fav_count }}</span>
