+++ favit/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import patterns, url
 
 
-urlpatterns = patterns('favorite.views',
+urlpatterns = patterns('favit.views',
     url(r'^add-or-remove$', 'add_or_remove'),
 )
